+++ src/main/java/com/climbing/global/jwt/JwtService.java
@@ -1,127 +1,32 @@
 package com.climbing.global.jwt;
 
-import com.auth0.jwt.JWT;
-import com.auth0.jwt.algorithms.Algorithm;
-import com.climbing.domain.member.repository.MemberRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Service;
 
-import java.util.Date;
 import java.util.Optional;
 
-@RequiredArgsConstructor
-@Service
-@Getter
-@Slf4j
-public class JwtService {
+public interface JwtService {
+    String createAccessToken(String email);
 
-    @Value("${jwt.secretKey}")
-    private String secretKey;
+    String createRefreshToken();
 
-    @Value("${jwt.access.expiration}")
-    private Long accessTokenExpirationPeriod;
+    void updateRefreshToken(String mail, String refreshToken);
 
-    @Value("${jwt.refresh.expiration}")
-    private Long refreshTokenExpirationPeriod;
+    void destroyRefreshToken(String email);
 
-    @Value("${jwt.access.header}")
-    private String accessHeader;
+    void sendAccessTokenAndRefreshToken(HttpServletResponse response, String accessToken, String refreshToken);
 
-    @Value("${jwt.refresh.header}")
-    private String refreshHeader;
+    void sendAccessToken(HttpServletResponse response, String accessToken);
 
-    private static final String ACCESS_TOKEN_SUBJECT = "AccessToken";
-    private static final String REFRESH_TOKEN_SUBJECT = "RefreshToken";
-    private static final String EMAIL_CLAIM = "email";
-    private static final String BEARER = "Bearer ";
+    Optional<String> extractAccessToken(HttpServletRequest request);
 
-    private final MemberRepository memberRepository;
+    Optional<String> extractRefreshToken(HttpServletRequest request);
 
-    public String createAccessToken(String email) {
-        Date now = new Date();
-        return JWT.create()
-                .withSubject(ACCESS_TOKEN_SUBJECT)
-                .withExpiresAt(new Date(now.getTime() + accessTokenExpirationPeriod))
-                .withClaim(EMAIL_CLAIM, email)
-                .sign(Algorithm.HMAC512(secretKey));
-    }
+    Optional<String> extractEmail(String accessToken);
 
-    public String createRefreshToken() {
-        Date now = new Date();
-        return JWT.create()
-                .withSubject(REFRESH_TOKEN_SUBJECT)
-                .withExpiresAt(new Date(now.getTime() + refreshTokenExpirationPeriod))
-                .sign(Algorithm.HMAC512(secretKey));
-    }
+    void setAccessTokenHeader(HttpServletResponse response, String accessToken);
 
-    public void sendAccessToken(HttpServletResponse response, String accessToken) {
-        response.setStatus(HttpServletResponse.SC_OK);
+    void setRefreshTokenHeader(HttpServletResponse response, String refreshToken);
 
-        response.setHeader(accessHeader, accessToken);
-        log.info("Access Token : {}", accessToken);
-    }
-
-    public void sendAccessTokenAndRefreshToken(HttpServletResponse response, String accessToken, String refreshToken) {
-        response.setStatus(HttpServletResponse.SC_OK);
-
-        setAccessTokenHeader(response, accessToken);
-        setRefreshTokenHeader(response, refreshToken);
-        log.info("Access Token, RefreshToken Header Complete");
-    }
-
-    public Optional<String> extractAccessToken(HttpServletRequest request) {
-        return Optional.ofNullable(request.getHeader(accessHeader))
-                .filter(accessToken -> accessToken.startsWith(BEARER))
-                .map(accessToken -> accessToken.replace(BEARER, ""));
-    }
-
-    public Optional<String> extractRefreshToken(HttpServletRequest request) {
-        return Optional.ofNullable(request.getHeader(refreshHeader))
-                .filter(refreshToken -> refreshToken.startsWith(BEARER))
-                .map(refreshToken -> refreshToken.replace(BEARER, ""));
-    }
-
-    public Optional<String> extractEmail(String accessToken) {
-        try {
-            return Optional.ofNullable(JWT.require(Algorithm.HMAC512(secretKey))
-                    .build()
-                    .verify(accessToken)
-                    .getClaim(EMAIL_CLAIM)
-                    .asString());
-        } catch (Exception e) {
-            log.error("유효하지 않은 토큰입니다.");
-            return Optional.empty();
-        }
-    }
-
-    public void setAccessTokenHeader(HttpServletResponse response, String accessToken) {
-        response.setHeader(accessHeader, accessToken);
-    }
-
-    public void setRefreshTokenHeader(HttpServletResponse response, String refreshToken) {
-        response.setHeader(refreshHeader, refreshToken);
-    }
-
-    public void updateRefreshToken(String email, String refreshToken) {
-        memberRepository.findByEmail(email)
-                .ifPresentOrElse(
-                        member -> member.updateRefreshToken(refreshToken),
-                        () -> new Exception("회원이 존재하지 않습니다.")
-                );
-    }
-
-    public boolean isTokenValid(String token) {
-        try {
-            JWT.require(Algorithm.HMAC512(secretKey)).build().verify(token);
-            return true;
-        } catch (Exception e) {
-            log.error("유효하지 않은 토큰입니다. {}", e.getMessage());
-            return false;
-        }
-    }
+    boolean isTokenValid(String token);
 }
